7.2

+++ Luogu/Sample.cpp
@@ -1,9 +1,9 @@
-#include <bits/stdc++.h>
+#include "bits/stdc++.h"
 using namespace std;
 
 using ll = long long;
 using vi = vector<int>;
-using vll = vector<long long>;
+using vll = vector<ll>;
 
 int main()
 {
